requirements.txt
# EVM托管钱包系统依赖

# 后端框架
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 数据库
sqlalchemy>=2.0.0
alembic>=1.10.0
python-dotenv>=1.0.0

# 区块链交互
web3>=6.0.0
eth-account>=0.11.0
eth-tester>=0.9.0b1
eth-abi>=5.0.1
eth-typing>=5.0.0  # 升级到与web3兼容的版本
eth-keys>=0.4.0
cryptography>=40.0.0

# AI对话控制
langchain>=0.0.267
langchain-core>=0.0.10
langchain-community>=0.0.10
langchain-experimental>=0.0.10
langchain-openai>=0.0.2
langsmith>=0.0.30
langserve>=0.0.30
langchainhub>=0.1.13
langchain-text-splitters>=0.0.1
langchain-chroma>=0.0.1
chroma-hnswlib>=0.7.3
openai>=1.3.0

# MCP协议
websockets>=11.0.0

# 测试
pytest>=7.0.0
httpx>=0.24.0